Order dependencies for release
When library A depends on B, B must be released first, so that its
`deps.edn` can reference an existing version on Clojars (maven checks &
fetches it to find out about dependencies).

This change creates a DAG (directed acyclic graph) of all the
dependencies included in the release and orders them topologically. That
way non-dependent libraries get released first. And the ones depended on
get released before the ones that depend on them

> Alternatively the deps could be released to a local repository (`~/.m2`) but that could be confusing on CI