Allow libraries to add custom CircleCI steps
Some libraries (like `shared/cljs-ipfs-http-client`) need extra steps
for their tests to run (like starting IPFS node). This change allows the
libraries to place `circleci-extra-steps.yml` under the library root
(not monorepo root, as its library specific). That way when that library
is determined to be tested, built & released, the contents of this file
gets added to the generated CircleCI config as a first thing for the
actions for that library.